Add tests demonstrating jackson serialization bug.
`GraphQLClientJacksonSerializer` has a bug related to is-prefixed
variable input fields. If the is-prefixed field is a `Boolean` field
it serializes correctly, but if the field type is not `Boolean`, then
the Jackson serializer silently omits the field entirely.

I added a similar test for `GraphQLClientKotlinXSerializer` to demonstrate
that the same bug does not exist on that serializer.